+++ packages/gatsby/src/utils/flags.ts
@@ -1,12 +1,44 @@
+import _ from "lodash"
+import semver from "semver"
+
 // Does this experiment run for only builds
 type executingCommand = "build" | "develop" | "all"
+
+export const satisfiesSemvers = (
+  semverConstraints: Record<string, string>
+): boolean => {
+  // Check each semver check for the flag.
+  // If any are false, then the flag doesn't pass
+  const result = _.toPairs(semverConstraints).every(
+    ([packageName, semverConstraint]) => {
+      let packageVersion: string
+      try {
+        packageVersion = require(`${packageName}/package.json`).version
+      } catch (e) {
+        return false
+      }
+
+      // We care if the semver check doesn't pass.
+      return semver.satisfies(packageVersion, semverConstraint, {
+        includePrerelease: true,
+      })
+    }
+  )
+
+  return result
+}
+
+export type fitnessEnum = true | false | "OPT_IN" | "LOCKED_IN"
 
 export interface IFlag {
   name: string
   env: string
   description: string
   command: executingCommand
-  telemetryId: string
+  /**
+   * Use string identifier to track enabled flag or false to disable any tracking (useful when flag becomes new defaults)
+   */
+  telemetryId: string | false
   // Heuristics for deciding if a flag is experimental:
   // - there are known bugs most people will encounter and that block being
   // able to use Gatsby normally
@@ -17,9 +49,24 @@ export interface IFlag {
   // resolved and ~50+ people have tested it, experimental should be set to
   // false.
   experimental: boolean
+  /**
+   * True means conditions for the feature are met and can be opted in by user.
+   *
+   * False means it'll be disabled despite the user setting it true e.g.
+   * it just won't work e.g. it doesn't have new enough version for something.
+   *
+   * OPT_IN means the gatsby will enable the flag (unless the user explicitly
+   * disables it.
+   *
+   * LOCKED_IN means that feature is enabled always (unless `noCI` condition is met).
+   * This is mostly to provide more meaningful terminal messages instead of removing
+   * flag from the flag list when users has the flag set in configuration
+   * (avoids showing unknown flag message and shows "no longer needed" message).
+   */
+  testFitness: (flag: IFlag) => fitnessEnum
   includedFlags?: Array<string>
   umbrellaIssue?: string
-  noCi?: boolean
+  noCI?: boolean
 }
 
 const activeFlags: Array<IFlag> = [
@@ -32,11 +79,10 @@ const activeFlags: Array<IFlag> = [
     description: `Enable all experiments aimed at improving develop server start time`,
     includedFlags: [
       `DEV_SSR`,
-      `QUERY_ON_DEMAND`,
-      `LAZY_IMAGES`,
       `PRESERVE_FILE_DOWNLOAD_CACHE`,
       `PRESERVE_WEBPACK_CACHE`,
     ],
+    testFitness: (): fitnessEnum => true,
   },
   {
     name: `DEV_SSR`,
@@ -45,26 +91,42 @@ const activeFlags: Array<IFlag> = [
     telemetryId: `DevSsr`,
     experimental: false,
     description: `SSR pages on full reloads during develop. Helps you detect SSR bugs and fix them without needing to do full builds.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/28138`,
+    umbrellaIssue: `https://gatsby.dev/dev-ssr-feedback`,
+    testFitness: (): fitnessEnum => true,
   },
   {
     name: `QUERY_ON_DEMAND`,
     env: `GATSBY_EXPERIMENTAL_QUERY_ON_DEMAND`,
     command: `develop`,
-    telemetryId: `QueryOnDemand`,
+    telemetryId: false,
     experimental: false,
     description: `Only run queries when needed instead of running all queries upfront. Speeds starting the develop server.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/27620`,
-    noCi: true,
+    umbrellaIssue: `https://gatsby.dev/query-on-demand-feedback`,
+    noCI: true,
+    testFitness: (): fitnessEnum => `LOCKED_IN`,
   },
   {
     name: `LAZY_IMAGES`,
     env: `GATSBY_EXPERIMENTAL_LAZY_IMAGES`,
     command: `develop`,
-    telemetryId: `LazyImageProcessing`,
-    experimental: true,
+    telemetryId: false,
+    experimental: false,
     description: `Don't process images during development until they're requested from the browser. Speeds starting the develop server. Requires gatsby-plugin-sharp@2.10.0 or above.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/27603`,
+    umbrellaIssue: `https://gatsby.dev/lazy-images-feedback`,
+    noCI: true,
+    testFitness: (): fitnessEnum => {
+      const semverConstraints = {
+        // Because of this, this flag will never show up
+        "gatsby-plugin-sharp": `>=2.10.0`,
+      }
+      if (satisfiesSemvers(semverConstraints)) {
+        return `LOCKED_IN`
+      } else {
+        // gatsby-plugin-sharp is either not installed or not new enough so
+        // just disable — it won't work anyways.
+        return false
+      }
+    },
   },
   {
     name: `PRESERVE_WEBPACK_CACHE`,
@@ -73,7 +135,8 @@ const activeFlags: Array<IFlag> = [
     telemetryId: `PreserveWebpackCache`,
     experimental: false,
     description: `Don't delete webpack's cache when changing gatsby-node.js & gatsby-config.js files.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/28331`,
+    umbrellaIssue: `https://gatsby.dev/cache-clearing-feedback`,
+    testFitness: (): fitnessEnum => true,
   },
   {
     name: `PRESERVE_FILE_DOWNLOAD_CACHE`,
@@ -82,7 +145,8 @@ const activeFlags: Array<IFlag> = [
     telemetryId: `PreserveFileDownloadCache`,
     experimental: false,
     description: `Don't delete the downloaded files cache when changing gatsby-node.js & gatsby-config.js files.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/28331`,
+    umbrellaIssue: `https://gatsby.dev/cache-clearing-feedback`,
+    testFitness: (): fitnessEnum => true,
   },
   {
     name: `FAST_REFRESH`,
@@ -91,7 +155,8 @@ const activeFlags: Array<IFlag> = [
     telemetryId: `FastRefresh`,
     experimental: false,
     description: `Use React Fast Refresh instead of the legacy react-hot-loader for instantaneous feedback in your development server. Recommended for versions of React >= 17.0.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/28390`,
+    umbrellaIssue: `https://gatsby.dev/fast-refresh-feedback`,
+    testFitness: (): fitnessEnum => true,
   },
   {
     name: `PARALLEL_SOURCING`,
@@ -100,7 +165,8 @@ const activeFlags: Array<IFlag> = [
     telemetryId: `ParallelSourcing`,
     experimental: true,
     description: `Run all source plugins at the same time instead of serially. For sites with multiple source plugins, this can speedup sourcing and transforming considerably.`,
-    umbrellaIssue: `https://github.com/gatsbyjs/gatsby/discussions/28336`,
+    umbrellaIssue: `https://gatsby.dev/parallel-sourcing-feedback`,
+    testFitness: (): fitnessEnum => true,
   },
 ]