+++ packages/gatsby/src/utils/webpack-utils.ts
@@ -515,7 +515,7 @@ export const createWebpackUtils = (
   rules.images = (): RuleSetRule => {
     return {
       use: [loaders.url()],
-      test: /\.(ico|svg|jpg|jpeg|png|gif|webp)(\?.*)?$/,
+      test: /\.(ico|svg|jpg|jpeg|png|gif|webp|avif)(\?.*)?$/,
     }
   }
 
@@ -727,7 +727,7 @@ export const createWebpackUtils = (
   }
 }
 
-function reactHasJsxRuntime(): boolean {
+export function reactHasJsxRuntime(): boolean {
   // We've got some complains about the ecosystem not being ready for automatic so we disable it by default.
   // People can use a custom babelrc file to support it
   // try {